# services/torrent-engine/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Torrent Stream API",
    "version": "0.1.0"
  },
  "paths": {
    "/torrents": {
      "get": {
        "summary": "List torrents",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "completed", "stopped", "all"],
              "default": "all"
            }
          },
          {
            "name": "view",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["summary", "full"],
              "default": "summary"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1 }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": { "type": "integer", "minimum": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/TorrentListSummary" },
                    { "$ref": "#/components/schemas/TorrentListFull" }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create torrent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTorrentJSON"
              },
              "examples": {
                "magnet": {
                  "value": {
                    "magnet": "magnet:?xt=urn:btih:..."
                  }
                }
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateTorrentMultipart"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentRecord"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported media type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/torrents/{id}": {
      "get": {
        "summary": "Get torrent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentRecord"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete torrent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "deleteFiles",
            "in": "query",
            "schema": { "type": "boolean", "default": false }
          }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/settings/hls": {
      "get": {
        "summary": "Get HLS settings",
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HLSSettings" } } } },
          "501": { "description": "Not configured", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "patch": {
        "summary": "Update HLS settings",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HLSSettings" } } }
        },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/HLSSettings" } } } },
          "400": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Update failed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/watch-history": {
      "get": {
        "summary": "Get recent watch history",
        "parameters": [
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1 } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/WatchPosition" } }
              }
            }
          },
          "400": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "501": { "description": "Not configured", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateTorrentJSON": {
        "type": "object",
        "properties": {
          "magnet": { "type": "string" },
          "name": { "type": "string" }
        },
        "required": ["magnet"]
      },
      "CreateTorrentMultipart": {
        "type": "object",
        "properties": {
          "torrent": { "type": "string", "format": "binary" },
          "name": { "type": "string" }
        },
        "required": ["torrent"]
      },
      "TorrentRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "status": { "type": "string" },
          "infoHash": { "type": "string" },
          "files": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FileRef" }
          },
          "mediaOrganization": { "$ref": "#/components/schemas/MediaOrganization" },
          "totalBytes": { "type": "integer", "format": "int64" },
          "doneBytes": { "type": "integer", "format": "int64" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "status"]
      },
      "TorrentSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "status": { "type": "string" },
          "progress": { "type": "number", "format": "double" },
          "doneBytes": { "type": "integer", "format": "int64" },
          "totalBytes": { "type": "integer", "format": "int64" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "TorrentListSummary": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TorrentSummary" }
          },
          "count": { "type": "integer" }
        }
      },
      "TorrentListFull": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TorrentRecord" }
          },
          "count": { "type": "integer" }
        }
      },
      "MediaOrganization": {
        "type": "object",
        "properties": {
          "contentType": { "type": "string", "enum": ["series", "movie", "mixed", "unknown"] },
          "groups": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MediaOrganizationGroup" }
          }
        }
      },
      "MediaOrganizationGroup": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string", "enum": ["series", "movie", "other"] },
          "title": { "type": "string" },
          "season": { "type": "integer" },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/MediaOrganizationItem" }
          }
        }
      },
      "MediaOrganizationItem": {
        "type": "object",
        "properties": {
          "fileIndex": { "type": "integer" },
          "filePath": { "type": "string" },
          "displayName": { "type": "string" },
          "season": { "type": "integer" },
          "episode": { "type": "integer" },
          "year": { "type": "integer" }
        }
      },
      "FileRef": {
        "type": "object",
        "properties": {
          "index": { "type": "integer" },
          "path": { "type": "string" },
          "length": { "type": "integer", "format": "int64" }
        }
      },
...
      "HLSSettings": {
        "type": "object",
        "properties": {
          "segmentDuration": { "type": "integer", "minimum": 2, "maximum": 10 },
          "ramBufSizeMB": { "type": "integer", "minimum": 4 },
          "prebufferMB": { "type": "integer", "minimum": 1 },
          "windowBeforeMB": { "type": "integer", "minimum": 1 },
          "windowAfterMB": { "type": "integer", "minimum": 4 }
        }
      },
...
      "WatchPosition": {
        "type": "object",
        "properties": {
          "torrentId": { "type": "string" },
          "fileIndex": { "type": "integer" },
          "position": { "type": "number", "format": "double" },
          "duration": { "type": "number", "format": "double" },
          "torrentName": { "type": "string" },
          "filePath": { "type": "string" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      }
    }
  }
}